almost all

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: kmumm <kmumm@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/04 17:00:32 by kmumm             #+#    #+#             */
-/*   Updated: 2021/10/04 17:34:33 by kmumm            ###   ########.fr       */
+/*   Updated: 2021/10/13 14:33:15 by kmumm            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned long	ft_strlen(const char *s)
+UINT	ft_strlen(const char *s)
 {
-	unsigned long	length;
+	UINT	length;
 
 	length = 0;
 	while (s[length])
